[transport] Fix transport to not require persistent data.
The shared memory transport will now copy the data into a temporary
buffer on a send. The receiver is passed a pointer to the buffer, and
it is his responsibility to delete it. This matches the semantics of
the MPI transport layer. These semantics weren't being followed in the
MCP, however, so it required touching those files as well.