Fixed a design bug --------------------
Found a bug caused by the handling of "SHARED_MEM_RESPONSE" packet by the user thread. The bug comes because of a race between the user and network threads. So, I moved this functionality into the network thread.
This fixed the problem to a certain extent but there are still problems apparently due to bad clock updations somewhere. The "timestamps" are not multiples of 10 as expected. I temporarily fixed this problem by modifying the 'netPullFromTransport' function in 'network.cc'. This modification can be removed after the bug in 'timestamps' is fixed.