[lcp][mcp] Move comm id handling to the LCP. Use LCP communication mechanisms where appropriate instead of hijacking MCPs (or other cores).
[network] Re-work NetPacket to be responsible for making data buffers
out of itself, replacing netExPacket and netCreateBuf. Also put
timestamp in NetPacket, eliminating NetQueueEntry.

Notes: Seems to avoid NaN issue on cannon. If this is true, I suspect
it is a consequence of the network changes. Also does not work for
distributed tests (globalSend in core_manager.cc fails), but passes
the single-process portion of regress_quick.