[Multi-process simulation] Made some (temporary) changes to the thread-spawning and initialization mechanism to allow multi-process simulation
This scheme is only supposed to be a temporary "proof of cocept" implementation, and will only work correctly for two processes as of now. Thread exiting is not correctly implemented, and one needs to end the process manually. All our regression tests except 'cannon' successfully work with 2-process simulation with this scheme