[pthread support] Bug fixes, assertions, and a pthreads_unit_test.
Added "-u CarbonSpawnThreadSpawner" to the link flags to fix a bug
where that symbol is undefined if CarbonStart/StopSim() are not included.
Added a pthreads unit test that runs through barrier and mutexes.
Added printout warnings for when attributes are non-null.
Added assertion for CarbonSpawnThreadSpawner being non-null in thread_start.