[M4 macros] Changed G_MALLOC etc to use malloc
These used to use valloc before.
This was the reason for all the ridiculously high addresses seen earlier.
This was happening because we don't have a #include <malloc.h> in MAIN_ENV,
which is where valloc is declared. Without a declaration, gcc assumes that
valloc returns an int, which gets us into trouble when running 64-bit