[pthreads][condition variable] Added a pthreads unit test that uses condition variables
Uses only pthread_cond_wait() and pthread_cond_broadcast() for now.
pthread_cond_signal() does not work yet since it causes the futex
syscall to be called with the FUTEX_WAKE_OP operation which is not
documented in the manpages. None of the Splash2 benchmarks call the
futex syscall with the FUTEX_WAKE_OP operation.